Finished commenting the new printer structure including the columns. Minor simplifications.
