+++ jsprit-core/src/main/java/com/graphhopper/jsprit/core/reporting/route/ActivityCostPrinterColumn.java
@@ -4,12 +4,24 @@ import java.util.function.Consumer;
 
 import com.graphhopper.jsprit.core.reporting.columndefinition.ColumnDefinition;
 
+/**
+ * Cost of the activity.
+ *
+ * @author balage
+ *
+ */
 public class ActivityCostPrinterColumn extends AbstractCostPrinterColumn {
 
+    /**
+     * Constructor.
+     */
     public ActivityCostPrinterColumn() {
         super();
     }
 
+    /**
+     * Constructor with a post creation decorator provided.
+     */
     public ActivityCostPrinterColumn(Consumer<ColumnDefinition.Builder> decorator) {
         super(decorator);
     }
