+++ jsprit-core/src/main/java/com/graphhopper/jsprit/core/reporting/route/EndTimePrinterColumn.java
@@ -6,12 +6,35 @@ import com.graphhopper.jsprit.core.problem.solution.route.activity.End;
 import com.graphhopper.jsprit.core.problem.solution.route.activity.TourActivity;
 import com.graphhopper.jsprit.core.reporting.columndefinition.ColumnDefinition;
 
+/**
+ * End time of the activity.
+ * <p>
+ * For route end the value is undefined (null), for other activities, it is the
+ * time when the activity is finished and the vehicle could progress toward the
+ * next activity.
+ * </p>
+ *
+ * @author balage
+ *
+ * @see {@linkplain ArrivalTimePrinterColumn}
+ * @see {@linkplain StartTimePrinterColumn}
+ * @see {@linkplain TravelDurationPrinterColumn}
+ * @see {@linkplain WaitingDurationPrinterColumn}
+ * @see {@linkplain OperationDurationPrinterColumn}
+ * @see {@linkplain ActivityDurationPrinterColumn}
+ */
 public class EndTimePrinterColumn extends AbstractTimePrinterColumn<EndTimePrinterColumn> {
 
+    /**
+     * Constructor.
+     */
     public EndTimePrinterColumn() {
         super();
     }
 
+    /**
+     * Constructor with a post creation decorator provided.
+     */
     public EndTimePrinterColumn(Consumer<ColumnDefinition.Builder> decorator) {
         super(decorator);
     }
