--- jsprit-core/src/main/java/com/graphhopper/jsprit/core/reporting/route/HumanReadableEnabled.java
@@ -1,7 +1,0 @@
-package com.graphhopper.jsprit.core.reporting.route;
-
-public interface HumanReadableEnabled<T extends HumanReadableEnabled<T>> {
-    public T withFormatter(HumanReadableTimeFormatter formatter);
-
-    public T asHumanReadable();
-}
