+++ jsprit-core/src/main/java/com/graphhopper/jsprit/core/reporting/route/JobPriorityPrinterColumn.java
@@ -10,14 +10,31 @@ import com.graphhopper.jsprit.core.reporting.columndefinition.ColumnAlignment;
 import com.graphhopper.jsprit.core.reporting.columndefinition.ColumnDefinition;
 import com.graphhopper.jsprit.core.reporting.columndefinition.StringColumnType;
 
+/**
+ * Priority of the job.
+ *
+ * <p>
+ * This column provides the named (LOW, MEDIUM, HIGH) representation of
+ * {@linkplain Job#getPriority()} of the associated job of the activity for job
+ * activities and null for other route activities.
+ * </p>
+ *
+ * @author balage
+ */
 public class JobPriorityPrinterColumn extends AbstractPrinterColumn<RoutePrinterContext, String, JobPriorityPrinterColumn> {
 
     private static final String[] PRIORITY_NAMES = new String[] { "", "HIGH", "MEDIUM", "LOW" };
 
+    /**
+     * Constructor.
+     */
     public JobPriorityPrinterColumn() {
         super();
     }
 
+    /**
+     * Constructor with a post creation decorator provided.
+     */
     public JobPriorityPrinterColumn(Consumer<ColumnDefinition.Builder> decorator) {
         super(decorator);
     }
