+++ jsprit-core/src/main/java/com/graphhopper/jsprit/core/reporting/route/SelectedTimeWindowPrinterColumn.java
@@ -10,12 +10,32 @@ import com.graphhopper.jsprit.core.problem.solution.route.activity.TimeWindow;
 import com.graphhopper.jsprit.core.problem.solution.route.activity.TourActivity;
 import com.graphhopper.jsprit.core.reporting.columndefinition.ColumnDefinition;
 
+/**
+ * The time window used in activity.
+ *
+ * <p>
+ * This is the time window which was choosen by the algorithm. The start time of
+ * the activity is within this time window and the end time is within or matches
+ * the end value of this time window.
+ * </p>
+ *
+ * @author balage
+ *
+ * @see {@linkplain TimeWindowsPrinterColumn}
+ * @see {@linkplain StartTimePrinterColumn}
+ */
 public class SelectedTimeWindowPrinterColumn extends AbstractTimeWindowPrinterColumn<SelectedTimeWindowPrinterColumn> {
 
+    /**
+     * Constructor.
+     */
     public SelectedTimeWindowPrinterColumn() {
         super();
     }
 
+    /**
+     * Constructor with a post creation decorator provided.
+     */
     public SelectedTimeWindowPrinterColumn(Consumer<ColumnDefinition.Builder> decorator) {
         super(decorator);
     }
@@ -25,6 +45,14 @@ public class SelectedTimeWindowPrinterColumn extends AbstractTimeWindowPrinterCo
         return "selTimeWindow";
     }
 
+    /**
+     * {@inheritDoc}
+     *
+     * <p>
+     * This implementation returns at most one time window: the one the activity
+     * start time is within.
+     * </p>
+     */
     @Override
     protected Collection<TimeWindow> getValue(RoutePrinterContext context) {
         TourActivity act = context.getActivity();
