+++ jsprit-core/src/main/java/com/graphhopper/jsprit/core/reporting/route/TransportCostPrinterColumn.java
@@ -6,14 +6,34 @@ import com.graphhopper.jsprit.core.problem.solution.route.activity.Start;
 import com.graphhopper.jsprit.core.problem.solution.route.activity.TourActivity;
 import com.graphhopper.jsprit.core.reporting.columndefinition.ColumnDefinition;
 
+/**
+ * The cost of travelling to the activity.
+ *
+ * <p>
+ * This is the cost of the transport from the previous to this activity. For the
+ * start of the route this value is undefined (null).
+ * </p>
+ * <p>
+ * This column is stateful and stores the previous activity.
+ * </p>
+ *
+ * @author balage
+ */
 public class TransportCostPrinterColumn extends AbstractCostPrinterColumn {
 
+    // The previous activity
     private TourActivity prevAct;
 
+    /**
+     * Constructor.
+     */
     public TransportCostPrinterColumn() {
         super();
     }
 
+    /**
+     * Constructor with a post creation decorator provided.
+     */
     public TransportCostPrinterColumn(Consumer<ColumnDefinition.Builder> decorator) {
         super(decorator);
     }
